handle compiler or lint error
