+++ store/datastore/ddl/sqlite3/11.sql
@@ -1,0 +1,8 @@
+-- +migrate Up
+
+ALTER TABLE builds ADD COLUMN build_error TEXT;
+UPDATE builds SET build_error = '';
+
+-- +migrate Down
+
+ALTER TABLE builds DROP COLUMN build_error;
